+++ src/main/java/com/DevEx/DevExBE/domain/handcarry/dto/HandcarryRequestDto.java
@@ -19,7 +19,7 @@ public class HandcarryRequestDto {
 
     private String endPoint;
 
-    private Long unitCosts;
+    private Float unitCosts;
 
     private Long maxWeight;
